+++ Changes.md
@@ -17,6 +17,7 @@ Fixes
 
 - NodeEditor, Viewer : Fixed bug that could cause unnecessary widget updates. In particular, this fixes flickering in the Viewer toolbar widgets when viewing the output of an InteractiveRender.
 - Viewer : Improved error handling when unable to create the requested renderer.
+- Inference : GPU device memory is now freed where possible after each computation.
 
 1.6.4.0 (relative to 1.6.3.0)
 =======